antlr_py/pyproject.toml
[tool.poetry]
name = "antlr_tutorial"
version = "0.1.0"
description = ""
authors = ["Ali Ghahraei Figueroa <ali.ghahraei@spectrumeffect.com>"]

[tool.poetry.dependencies]
python = "^3.8"
antlr4-python3-runtime = "^4.8"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
